Completed 08_enums excercise

+++ exercises/08_enums/enums3.rs
@@ -6,7 +6,7 @@
 // hint.
 enum Message {
     // TODO: implement the message variant types based on their usage below
-    ChangeColor(u32,u32,u32),
+    ChangeColor(u8,u8,u8),
     Echo(String),
     Move(Point),
     Quit,
